requirements.txt
requests
pandas
deltalake
python-dotenv
logging
